Remove clientOptions from keyvault common
