+++ sdk/keyvault/azure-security-keyvault-common/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## 4.0.0-beta.2 (Unreleased)
 
+### Breaking Changes
+
+- Removed `KeyVaultException`.
+- Removed `ClientOptions`.
 
 ## 4.0.0-beta.1 (2021-04-07)
