+++ sdk/keyvault/azure-security-keyvault-common/CMakeLists.txt
@@ -31,7 +31,6 @@ set(
     inc/azure/keyvault/common/internal/base64url.hpp
     inc/azure/keyvault/common/internal/keyvault_pipeline.hpp
     inc/azure/keyvault/common/internal/unix_time_helper.hpp
-    inc/azure/keyvault/common/client_options.hpp
     inc/azure/keyvault/common/keyvault_constants.hpp
     inc/azure/keyvault/common/keyvault_exception.hpp
     inc/azure/keyvault/common/version.hpp
